+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fixed OTLP gRPC exporter silently failing if scheme is not specified in endpoint.
   ([#1806](https://github.com/open-telemetry/opentelemetry-python/pull/1806))
+- Rename CompositeHTTPPropagator to CompositePropagator as per specification.
+  ([#1807](https://github.com/open-telemetry/opentelemetry-python/pull/1807))
 - Propagators use the root context as default for `extract` and do not modify
   the context if extracting from carrier does not work.
   ([#1811](https://github.com/open-telemetry/opentelemetry-python/pull/1811))